feat: useServer custom-hook 생성
