+++ src/components/templates/Page/index.js
@@ -1,58 +1,15 @@
-import Cookies from 'js-cookie';
 import { PropTypes } from 'prop-types';
-import React, { useEffect, useState } from 'react';
-import { useDispatch } from 'react-redux';
-import { useHistory } from 'react-router-dom';
+import React from 'react';
 
-import { ACCESS_TOKEN, ROUTE, SERVER_ID, SERVER_LIST } from '../../../constants';
-import { loginByToken } from '../../../redux/userSlice';
+import { SERVER_LIST } from '../../../constants';
+import useServer from '../../../hooks/useServer';
 import { NavBar, ServerSelect } from '../..';
 import { Header, Main, ServerSelectButton } from './style';
 
 export const Page = (props) => {
   const { children } = props;
 
-  const dispatch = useDispatch();
-  const history = useHistory();
-
-  const serverId = Cookies.get(SERVER_ID);
-  const accessToken = Cookies.get(ACCESS_TOKEN);
-
-  const [isServerSelectOpen, setIsServerSelectOpen] = useState(!serverId);
-
-  useEffect(() => {
-    if (!serverId) {
-      return;
-    }
-
-    if (!accessToken) {
-      history.push(ROUTE.LOGIN);
-      return;
-    }
-
-    dispatch(loginByToken({ accessToken }));
-  }, []);
-
-  const handleOpenModal = () => setIsServerSelectOpen(true);
-  const handleCloseModal = () => setIsServerSelectOpen(false);
-
-  const setNewServer = (serverId) => Cookies.set(SERVER_ID, serverId);
-
-  const handleServerSubmit = (e) => {
-    e.preventDefault();
-
-    const selectedId = e.target.serverSelect.value;
-
-    if (!selectedId) return;
-    if (selectedId === serverId) {
-      handleCloseModal();
-      return;
-    }
-
-    setNewServer(selectedId);
-    handleCloseModal();
-    history.push(ROUTE.LOGOUT);
-  };
+  const { handleOpenModal, handleServerSubmit, isServerSelectOpen, serverId } = useServer();
 
   return (
     <>
